[Common] create user exam service

+++ Edulingual.Domain/Entities/UserExam.cs
@@ -5,7 +5,7 @@ namespace Edulingual.Domain.Entities;
 public class UserExam : BaseEntity
 {
     public double Score { get; set; } = 0;
-    public int TotalQuestionRight { get; set; } = 0;
+    public double TotalQuestionRight { get; set; } = 0;
     public Guid UserId { get; set; }
     public Guid ExamId { get; set; }
